--- public/index.html
@@ -1,16 +1,0 @@
-<!-- <!DOCTYPE html> -->
-<html lang="ja">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title></title>
-</head>
-
-<body>
-  <div id="root"></div>
-</body>
-<script src="bundle.js" charset="utf-8"></script>
-
-</html>
