+++ src/client/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>React and Webpack4</title>
+    <title>TodoApp</title>
 </head>
 <body>
 <section id="index"></section>
